move.pathTo(x,z,d)  works!! default tether set to 34 blacklist defaults to: itemTypes.noMine
Notes on pathTo:
It has basic maze solving skills thus it will move backwards to try to
get forwards.  Thus the necessity to set a rational default tether, in the off chance
a dumb fuck trys to tell it to go to an unreachable point. It will give up with an error
and return to starting position but only after exhausting all available spapces it can reach :).

Odd bug, the necesity to always face forward after every move, fix this at your own peril it seems to break everything

directional preference.  you can set a preference on which direction is preffered.